Inject the robot identity instead of resolving it in the core
Resolving a credential means knowing about environment variables, config
files, a key file under someone's HOME, and POSIX permission rules.
Those are properties of the machine, not of the OTA protocol — and a
developer at a shell and a robot on a service account answer them
differently.

Leaving that in the core had a concrete consequence: a human running the
tool on a robot was attributed to *the robot*, because a key file
happened to be on disk. One command could even split across identities,
resolving a pinned archive as the user and downloading its packages as
the robot.

- `RobotIdentity` (api key, node key, client version) is part of the
  context. The core builds headers from it and asks no questions.
- Resolution moves to `commands/robot_credentials.py`, which the CLI
  uses to build an identity and hand it to `configure()`. About 260
  lines of environment and file probing leave the core.
- Refusing half a credential — a key with no node key — is resolution's
  job now, not the download path's. A robot request is scoped to a node, so
  half a credential would fail every request.

The environment no longer grants an identity: the caller decides who it
is. A machine with no credential simply makes no robot requests, which
is what a workstation should do.

212 tests. The CLI resolves and injects the same credential it always
used, so `raisin install` behaves as before.
